This is my final update for the night. The map initialization works, and you can claim a route by double-tapping it.
To do this, I had to add a bunch of functions to the IGamePresenter, IGameModel, Game object and GameMap object. This allows a player to send a request to claim a route, which travels all the way to the GameMap which updates the route, sending a reply back.

We'll need to clean this up for future phases, but it works. I updated the FakeModel to work for my needs, so you may need to make some changes to the ModelFacade for your test cases.

Everything seems to work, but it's 2:15 am and I need some sleep.

Also I combined our Color enums to make it work better for drawing routes. I also added manditory latitude/longitude values to the city constructor.